requirements.txt
PyMuPDF==1.24.5
python-docx==1.1.2
docx2txt==0.8
Pillow==10.3.0
